support for DB_USERNAME, DB_PASSWORD and DB_URL
