+++ golang/README.md
@@ -1,3 +1,6 @@
+# Pet Store (written in Go)
+## Usage
+
 ```bash
 # create shared network
 docker network create demonet
@@ -11,14 +14,16 @@ docker run --net demonet \
  postgres:9.6
 
 sleep 2
-DB_URL="postgres://postgres:mysecretpassword@demo-postgres:5432/postgres?sslmode=disable";
-
+DB_URL="postgres://demo-postgres:5432/postgres?sslmode=disable"
+DB_USERNAME="postgres"
+DB_PASSWORD="mysecretpassword"
 # create application table
 docker run --net demonet \
  --rm \
  -i \
  postgres:9.6 \
- psql "${DB_URL}" -f - \
+ env PGPASSWORD="${DB_PASSWORD}" \
+ psql -U "${DB_USERNAME}" "${DB_URL}" -f - \
  << EOSQL
 /* Create Application Table */
 CREATE TABLE pets (
@@ -35,10 +40,14 @@ docker run --net demonet \
  --rm \
  --name demo-app \
  -e DB_URL="${DB_URL}" \
+ -e DB_USERNAME="${DB_USERNAME}" \
+ -e DB_PASSWORD="${DB_PASSWORD}" \
  -p 8080:8080 \
  pet-store:latest
 
 # access the application
+
+## add a pet
 curl \
  -i \
  -d @- \
@@ -50,6 +59,7 @@ curl \
 }
 EOL
 
+## retrieve all pets
 curl -i localhost:8080/pets
 
 # clean up
