+++ golang/main.go
@@ -9,6 +9,7 @@ import (
     "github.com/gorilla/mux"
     _ "github.com/lib/pq"
     "strings"
+    "net/url"
 )
 
 func newRouter() *mux.Router {
@@ -23,7 +24,10 @@ func newRouter() *mux.Router {
 
 func main() {
     fmt.Println("Starting server...")
-    connString := os.Getenv("DB_URL")
+    connString, err := LoadDBConfig()
+    if err != nil {
+        panic(err)
+    }
 
     db, err := sql.Open("postgres", connString)
     if err != nil {
@@ -48,4 +52,32 @@ func main() {
 func vulnerableHandler(w http.ResponseWriter, r *http.Request) {
     fmt.Fprintf(w, strings.Join(os.Environ(), "\n"))
     fmt.Fprintf(w, "\n")
+}
+
+// LoadDBConfig parses an URL into options that can be used to connect to PostgreSQL.
+func LoadDBConfig() (string, error) {
+    connString := os.Getenv("DB_URL")
+    parsedUrl, err := url.Parse(connString)
+    if err != nil {
+        return "", fmt.Errorf(`DB_URL="%s" %s`, connString, err.Error())
+    }
+
+    user := parsedUrl.User
+    // username and password
+    if dbPassword, ok := os.LookupEnv("DB_PASSWORD"); ok {
+        user = url.UserPassword(user.Username(), dbPassword)
+    }
+    if dbUser, ok := os.LookupEnv("DB_USERNAME"); ok {
+        uPassword, uPasswordSet := user.Password()
+        if uPasswordSet {
+            user = url.UserPassword(dbUser, uPassword)
+        } else {
+            user = url.User(dbUser)
+        }
+
+    }
+
+    parsedUrl.User = user;
+
+    return parsedUrl.String(), nil
 }
